Make cache retention configurable via opts
